+++ udata_geoplatform/__init__.py
@@ -6,5 +6,5 @@ Connects uData with Geoplatform (geo.data.gouv.fr)
 '''
 from __future__ import unicode_literals
 
-__version__ = '1.0.1'
+__version__ = '1.0.2.dev'
 __description__ = 'Connects uData with Geoplatform (geo.data.gouv.fr)'