+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Migrate to python3 🐍 [#1](https://github.com/etalab/udata-geoplatform/pull/1)
 
+## 1.0.1 (2018-08-03)
+
+- Pass current locale to preview URL [#3](https://github.com/etalab/udata-geoplatform/pull/3)
+
 ## 1.0.0 (2018-06-06)
 
 - Initial release